# docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "HappyTailz API",
    "version": "1.0.0",
    "description": "Local Swagger documentation for HappyTailz backend"
  },
  "servers": [
    { "url": "http://localhost:3000", "description": "Local" },
    { "url": "https://happytailz-9qf9.onrender.com/", "description": "Render" }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Pet": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "ownerId": { "type": "string" },
          "name": { "type": "string" },
          "species": { "type": "string" },
          "breed": { "type": "string" },
          "age": { "type": "number" },
          "gender": { "type": "string" },
          "imageUrl": { "type": "string", "format": "uri", "description": "Direct URL of the pet image" }
        }
      },
...
    }
  },
...
  "paths": {
...
    "/api/auth/register": {
      "post": {
        "summary": "Register user",
        "tags": ["Auth"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "email", "password"],
                "properties": {
                  "name": { "type": "string" },
                  "email": { "type": "string", "format": "email" },
                  "password": { "type": "string", "minLength": 6 },
                  "role": { "type": "string", "enum": ["admin", "driver", "user"] }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Created" },
          "409": { "description": "Email exists" }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "summary": "Login",
        "tags": ["Auth"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": { "type": "string", "format": "email" },
                  "password": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "OK" }, "401": { "description": "Invalid" } }
      }
    },
    "/api/pets": {
      "get": {
        "summary": "List my pets",
        "tags": ["Pets"],
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "OK" } }
      },
      "post": {
        "summary": "Create pet",
        "tags": ["Pets"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["name", "species"],
                "properties": {
                  "name": { "type": "string" },
                  "species": { "type": "string" },
                  "breed": { "type": "string" },
                  "age": { "type": "number" },
                  "gender": { "type": "string" },
                  "image": { "type": "string", "format": "binary", "description": "Image file upload (field name: image)" },
                  "imageUrl": { "type": "string", "format": "uri", "description": "Optional direct image URL instead of file" }
                }
              }
            },
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "species"],
                "properties": {
                  "name": { "type": "string" },
                  "species": { "type": "string" },
                  "breed": { "type": "string" },
                  "age": { "type": "number" },
                  "gender": { "type": "string" },
                  "imageUrl": { "type": "string", "format": "uri" }
                }
              }
            }
          }
        },
        "responses": { "201": { "description": "Created" } }
      }
    },
    "/api/pets/{id}": {
      "parameters": [
        { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
      ],
      "put": {
        "summary": "Update my pet",
        "tags": ["Pets"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "species": { "type": "string" },
                  "breed": { "type": "string" },
                  "age": { "type": "number" },
                  "gender": { "type": "string" },
                  "image": { "type": "string", "format": "binary" },
                  "imageUrl": { "type": "string", "format": "uri" }
                }
              }
            },
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Pet" }
            }
          }
        },
        "responses": { "200": { "description": "OK" }, "404": { "description": "Not Found" } }
      },
      "delete": {
        "summary": "Delete my pet",
        "tags": ["Pets"],
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "Deleted" }, "404": { "description": "Not Found" } }
      }
    },
...
    "/api/admin/users": {
      "get": {
        "summary": "List all users",
        "tags": ["Admin - Users"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "role", "in": "query", "schema": { "type": "string", "enum": ["admin", "driver", "user"] } },
          { "name": "q", "in": "query", "schema": { "type": "string" }, "description": "Search by name/email" }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "_id": { "type": "string" },
                      "name": { "type": "string" },
                      "email": { "type": "string" },
                      "role": { "type": "string" },
                      "activePlan": { "type": "object", "nullable": true }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create user",
        "tags": ["Admin - Users"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["name", "email", "password"],
                "properties": {
                  "name": { "type": "string" },
                  "email": { "type": "string", "format": "email" },
                  "password": { "type": "string", "minLength": 6 },
                  "role": { "type": "string", "enum": ["admin", "driver", "user" ] },
                  "phone": { "type": "string" },
                  "address": { "type": "string" },
                  "image": { "type": "string", "format": "binary" },
                  "imageUrl": { "type": "string", "format": "uri" }
                }
              }
            },
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "email", "password"],
                "properties": {
                  "name": { "type": "string" },
                  "email": { "type": "string", "format": "email" },
                  "password": { "type": "string", "minLength": 6 },
                  "role": { "type": "string", "enum": ["admin", "driver", "user" ] },
                  "phone": { "type": "string" },
                  "address": { "type": "string" },
                  "imageUrl": { "type": "string", "format": "uri" }
                }
              }
            }
          }
        },
        "responses": { "201": { "description": "Created" }, "409": { "description": "Email exists" } }
      }
    },
    "/api/admin/users/{id}": {
      "parameters": [
        { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
      ],
      "get": {
        "summary": "Get user by id",
        "tags": ["Admin - Users"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "_id": { "type": "string" },
                    "name": { "type": "string" },
                    "email": { "type": "string" },
                    "role": { "type": "string" },
                    "activePlan": { "type": "object", "nullable": true }
                  }
                }
              }
            }
          },
          "404": { "description": "Not Found" }
        }
      },
      "put": {
        "summary": "Update user",
        "tags": ["Admin - Users"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "email": { "type": "string", "format": "email" },
                  "password": { "type": "string", "minLength": 6 },
                  "role": { "type": "string", "enum": ["admin", "driver", "user"] },
                  "phone": { "type": "string" },
                  "address": { "type": "string" },
                  "image": { "type": "string", "format": "binary" },
                  "imageUrl": { "type": "string", "format": "uri" }
                }
              }
            },
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "email": { "type": "string", "format": "email" },
                  "password": { "type": "string", "minLength": 6 },
                  "role": { "type": "string", "enum": ["admin", "driver", "user"] },
                  "phone": { "type": "string" },
                  "address": { "type": "string" },
                  "imageUrl": { "type": "string", "format": "uri" }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "OK" }, "404": { "description": "Not Found" } }
      },
      "delete": { "summary": "Delete user", "tags": ["Admin - Users"], "security": [{ "bearerAuth": [] }], "responses": { "200": { "description": "Deleted" }, "404": { "description": "Not Found" } } }
    },
...
  }
}